Use doubles over floats, fix int rounding bug in Y position

+++ src/player.h
@@ -4,13 +4,13 @@
 class Player
 {
 public:
-  Player() : x(0), y(-100), width(46), height(94) {}
-  float x;
-  float y;
-  float width;
-  float height;
+  Player() : x(0.0f), y(-100.0f), width(46.0f), height(94.0f) {}
+  double x;
+  double y;
+  double width;
+  double height;
   bool isGrounded;
-  float yAccel;
+  double yAccel;
 };
 
 #endif
